+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.alexparra.chatapp">
+    package="com.psandroidlabs.chatapp">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_WIFI_STATE" />
@@ -8,24 +8,24 @@
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE" />
 
     <application
-        android:name=".MainApplication"
+        android:name="com.psandroidlabs.chatapp.MainApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Socket">
-        <service android:name=".ServerService">
+        <service android:name="com.psandroidlabs.chatapp.ServerService">
             <intent-filter>
                 <action android:name="android.intent.action.STOP" />
             </intent-filter>
         </service>
         <receiver
-            android:name=".ActionManager"
+            android:name="com.psandroidlabs.chatapp.ActionManager"
             android:exported="true" />
 
         <activity
-            android:name="com.alexparra.chatapp.MainActivity"
+            android:name="com.psandroidlabs.chatapp.MainActivity"
             android:launchMode="singleTask">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
