--- app/src/main/java/com/alexparra/chatapp/MainActivity.kt
@@ -1,12 +1,0 @@
-package com.alexparra.chatapp
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-
-class MainActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-    }
-}
