--- app/src/main/java/com/alexparra/chatapp/MainApplication.kt
@@ -1,19 +1,0 @@
-package com.alexparra.chatapp
-
-import android.app.Application
-import android.content.Context
-
-class MainApplication : Application() {
-
-    init {
-        instance = this
-    }
-
-    companion object {
-        private lateinit var instance: Application
-
-        fun applicationContext(): Context {
-            return instance.applicationContext
-        }
-    }
-}
