--- app/src/main/java/com/alexparra/chatapp/models/Enums.kt
@@ -1,5 +1,0 @@
-package com.alexparra.chatapp.models
-
-enum class UserType {
-    CLIENT, SERVER
-}
