--- app/src/main/java/com/alexparra/chatapp/models/Message.kt
@@ -1,13 +1,0 @@
-package com.alexparra.chatapp.models
-
-// TODO CREATE TO STRING METHOD TO TRANSFORM INTO SENDABLE MESSAGE
-data class Message(
-    val type: MessageType,
-    val username: String,
-    val message: String,
-    val time: String
-)
-
-enum class MessageType {
-    SENT, RECEIVED, JOINED, ATTENTION
-}
