--- app/src/main/java/com/alexparra/chatapp/utils/Constants.kt
@@ -1,5 +1,0 @@
-package com.alexparra.chatapp.utils
-
-object Constants {
-    const val ACTION_STOP = "STOP"
-}
