+++ app/src/main/java/com/psandroidlabs/chatapp/ActionManager.kt
@@ -1,0 +1,18 @@
+package com.psandroidlabs.chatapp
+
+import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
+import androidx.localbroadcastmanager.content.LocalBroadcastManager
+import com.psandroidlabs.chatapp.utils.Constants
+
+class ActionManager : BroadcastReceiver() {
+    override fun onReceive(context: Context?, intent: Intent?) {
+        if (context == null) {
+            return
+        }
+
+        val broadcastIntent = Intent(Constants.ACTION_STOP)
+        LocalBroadcastManager.getInstance(context).sendBroadcast(broadcastIntent)
+    }
+}
