+++ app/src/main/java/com/psandroidlabs/chatapp/MainActivity.kt
@@ -1,0 +1,12 @@
+package com.psandroidlabs.chatapp
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class MainActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+    }
+}
