+++ app/src/main/java/com/psandroidlabs/chatapp/MainApplication.kt
@@ -1,0 +1,19 @@
+package com.psandroidlabs.chatapp
+
+import android.app.Application
+import android.content.Context
+
+class MainApplication : Application() {
+
+    init {
+        instance = this
+    }
+
+    companion object {
+        private lateinit var instance: Application
+
+        fun applicationContext(): Context {
+            return instance.applicationContext
+        }
+    }
+}
