+++ app/src/main/java/com/psandroidlabs/chatapp/models/Enums.kt
@@ -1,0 +1,5 @@
+package com.psandroidlabs.chatapp.models
+
+enum class UserType {
+    CLIENT, SERVER
+}
