+++ app/src/main/java/com/psandroidlabs/chatapp/utils/Constants.kt
@@ -1,0 +1,5 @@
+package com.psandroidlabs.chatapp.utils
+
+object Constants {
+    const val ACTION_STOP = "STOP"
+}
